Update code for headless execution
