--- config.json
@@ -1,6 +1,0 @@
-{
-    "email": "",
-    "password": "",
-    "tibiaLocation": "",
-    "resultsLocation": ""
-}
