+++ install.sh
@@ -5,3 +5,5 @@ sudo apt-get install -y python3-tk
 sudo apt-get install -y libpython3-dev
 sudo apt-get install -y tesseract-ocr
 sudo apt-get install -y tesseract-ocr-deu
+sudo apt-get install -y scrot
+sudo apt-get install -y xvfb xserver-xephyr tigervnc-standalone-server x11-utils gnumeric
