+++ src/items.csv
@@ -30,6 +30,7 @@ Afflicted Strider Head,36789
 Afflicted Strider Worms,36790
 Aggressive Fluid,14055
 Airtight Cloth,21870
+Albino Dragon Leather,44157
 Albino Plate,19358
 Alchemical Apparatus,8996
 Alchemist's Notepad,40594
@@ -641,6 +642,7 @@ Bottle of Tibian Wine,24948
 Bottle of Vinegar,31976
 Bottle of Whisper Beer,6106
 Bottle with Rat Urine,17671
+Bottled Dragon Breath,44179
 Bottled Yolk,40598
 Bow,3350
 Bow of Carving,23327
